requirements.txt
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.25.2
scipy==1.11.2
# Thư viện cho API của các mô hình AI
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
requests>=2.31.0
